refactor: moved filter state to its own model
