+++ changelogs/CHANGELOG - latest.md
@@ -1,2 +1,7 @@
 ## Changelog:
+- Added filter page to limit time period of dates to be shown
+- Some small style tweaks
 
+
+## Technical changes:
+-
